refactor: altera tipo de propriedades aceitas pelo zod

+++ src/interfaces/ICar.ts
@@ -2,8 +2,8 @@ import { z } from 'zod';
 import { VehicleZodSchema } from './IVehicle';
 
 export const CarZodSchema = z.object({
-  doorsQty: z.number().min(2).max(4),
-  seatsQty: z.number().min(2).max(7),
+  doorsQty: z.number().int().min(2).max(4),
+  seatsQty: z.number().int().min(2).max(7),
 }).merge(VehicleZodSchema);
 
 export type ICar = z.infer<typeof CarZodSchema>;
